refactor: map-api에 map 관련 api 통합
