+++ src/app/store.ts
@@ -15,14 +15,7 @@ import {
   storyApi,
 } from '../features/shop/shop-management';
 import { postApi } from '../features/post';
-import {
-  mapApi,
-  oneDayClassApi,
-  mapStoryApi,
-  favoriteApi,
-  reviewApi,
-  mapArtistApi,
-} from '../features/map';
+import { mapApi } from '../features/map';
 import { chatApi } from '../features/chat/api/chat-api';
 import { myPageApi } from '../features/my-page/api/my-page-api';
 import { classReservationApi } from '../features/shop/class-reservation-management/api/class-reservation-api';
@@ -51,11 +44,6 @@ export const store = configureStore({
     [storyApi.reducerPath]: storyApi.reducer,
     [businessHoursApi.reducerPath]: businessHoursApi.reducer,
     [mapApi.reducerPath]: mapApi.reducer,
-    [oneDayClassApi.reducerPath]: oneDayClassApi.reducer,
-    [mapStoryApi.reducerPath]: mapStoryApi.reducer,
-    [favoriteApi.reducerPath]: favoriteApi.reducer,
-    [reviewApi.reducerPath]: reviewApi.reducer,
-    [mapArtistApi.reducerPath]: mapArtistApi.reducer,
     [chatApi.reducerPath]: chatApi.reducer,
     [myPageApi.reducerPath]: myPageApi.reducer,
     [userMeApi.reducerPath]: userMeApi.reducer,
@@ -79,11 +67,6 @@ export const store = configureStore({
       .concat(storyApi.middleware)
       .concat(businessHoursApi.middleware)
       .concat(mapApi.middleware)
-      .concat(oneDayClassApi.middleware)
-      .concat(mapStoryApi.middleware)
-      .concat(favoriteApi.middleware)
-      .concat(reviewApi.middleware)
-      .concat(mapArtistApi.middleware)
       .concat(chatApi.middleware)
       .concat(myPageApi.middleware)
       .concat(userMeApi.middleware)
