+++ src/features/map/index.ts
@@ -5,8 +5,3 @@ export {
   useGetShopMarkersQuery,
   useGetShopDetailQuery,
 } from './api/map-api';
-export { oneDayClassApi } from './api/one-day-class-api';
-export { mapStoryApi } from './api/map-story-api';
-export { favoriteApi } from './api/favorite-api';
-export { reviewApi } from './api/review-api';
-export { mapArtistApi } from './api/map-artist-api';
